[REF] move coop_shift_state into coop_shift
