+++ louve_addons/coop_shift/data/ir_sequence.xml
