+++ louve_addons/coop_shift/model/__init__.py
@@ -32,5 +32,7 @@ from . import shift_type
 from . import shift_ticket
 from . import shift_template_ticket
 from . import shift_mail_registration
+from . import shift_extension
+from . import shift_extension_type
 from . import res_partner
 from . import product
