+++ louve_addons/coop_shift/views/view_shift_extension.xml
