+++ louve_addons/coop_shift/views/view_shift_extension_type.xml
