+++ louve_addons/coop_shift_state/__init__.py
@@ -1,2 +1,1 @@
 # -*- encoding: utf-8 -*-
-from . import models
