--- louve_addons/coop_shift_state/models/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import res_partner
-from . import shift_extension
-from . import shift_extension_type
