--- louve_addons/coop_shift_state/security/ir.model.access.csv
@@ -1,8 +1,0 @@
-id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-,,,,,,,
-,,,,,,,
-access_shift_extension_user,access_shift_extension_user,coop_shift_state.model_shift_extension,coop_shift.group_shift_user,1,0,0,0
-access_shift_extension_type_manager,access_shift_extension_type_manager,coop_shift_state.model_shift_extension,coop_shift.group_shift_manager,1,1,1,0
-,,,,,,,
-access_shift_extension_type_user,access_shift_extension_type_user,coop_shift_state.model_shift_extension_type,coop_shift.group_shift_user,1,0,0,0
-access_shift_extension_type_manager,access_shift_extension_type_manager,coop_shift_state.model_shift_extension_type,coop_shift.group_shift_manager,1,1,1,0
